+++ src/context/authContext.tsx
@@ -1,0 +1,52 @@
+import { createContext, useContext, useState } from "react";
+import { useMemo } from "../@lib";
+import { User } from "../types";
+import { useNotification } from "./notificationContext";
+
+interface AuthContextType {
+  user: User | null;
+  login: (email: string, password: string) => void;
+  logout: () => void;
+}
+
+const AuthContext = createContext<AuthContextType | undefined>(undefined);
+
+// Provider 별도 분리
+// App 에서 관리 시 전체 Provider 리렌더링 발생
+export function AuthProvider({ children }: { children: React.ReactNode }) {
+  const { addNotification } = useNotification();
+  const [user, setUser] = useState<User | null>(null);
+
+  const login = (email: string, _password: string) => {
+    setUser({ id: 1, name: "홍길동", email });
+    addNotification("성공적으로 로그인되었습니다", "success");
+  };
+
+  const logout = () => {
+    setUser(null);
+    addNotification("로그아웃되었습니다", "info");
+  };
+
+  const authValue = useMemo(
+    () => ({
+      user,
+      login,
+      logout,
+    }),
+    [user]
+  );
+
+  return (
+    <AuthContext.Provider value={authValue}>{children}</AuthContext.Provider>
+  );
+}
+
+export function useAuth() {
+  const context = useContext(AuthContext);
+  if (context === undefined) {
+    throw new Error("useAuth는 AuthProvider 내부에서 사용해야 합니다.");
+  }
+  return context;
+}
+
+export { AuthContext };
