+++ src/context/notificationContext.tsx
@@ -1,0 +1,63 @@
+import { createContext, useContext, useState } from "react";
+import { useMemo } from "../@lib";
+import { Notification } from "../types";
+
+interface NotificationContextType {
+  notifications: Notification[];
+  addNotification: (message: string, type: Notification["type"]) => void;
+  removeNotification: (id: number) => void;
+}
+
+export const NotificationContext = createContext<
+  NotificationContextType | undefined
+>(undefined);
+
+// Provider 별도 분리
+// App 에서 관리 시 전체 Provider 리렌더링 발생
+export function NotificationProvider({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  const [notifications, setNotifications] = useState<Notification[]>([]);
+
+  const addNotification = (message: string, type: Notification["type"]) => {
+    const newNotification: Notification = {
+      id: Date.now(),
+      message,
+      type,
+    };
+    setNotifications((prev) => [...prev, newNotification]);
+  };
+
+  const removeNotification = (id: number) => {
+    setNotifications((prev) =>
+      prev.filter((notification) => notification.id !== id)
+    );
+  };
+
+  const notificationValue = useMemo(
+    () => ({
+      notifications,
+      addNotification,
+      removeNotification,
+    }),
+    [notifications]
+  );
+
+  return (
+    <NotificationContext.Provider value={notificationValue}>
+      {children}
+    </NotificationContext.Provider>
+  );
+}
+
+export function useNotification() {
+  const context = useContext(NotificationContext);
+  if (context === undefined) {
+    throw new Error(
+      "useNotification은 NotificationProvider 내부에서 사용해야 합니다."
+    );
+  }
+  return context;
+}
