+++ src/context/themeContext.tsx
@@ -1,0 +1,37 @@
+import { createContext, useContext, useState } from "react";
+import { useMemo } from "../@lib";
+
+interface ThemeContextType {
+  theme: string;
+  toggleTheme: () => void;
+}
+
+const ThemeContext = createContext<ThemeContextType | undefined>(undefined);
+
+// Provider 별도 분리
+// App 에서 관리 시 전체 Provider 리렌더링 발생
+export function ThemeProvider({ children }: { children: React.ReactNode }) {
+  const [theme, setTheme] = useState("light");
+
+  // context value를 메모이제이션
+  const value = useMemo(
+    () => ({
+      theme,
+      toggleTheme: () =>
+        setTheme((prev) => (prev === "light" ? "dark" : "light")),
+    }),
+    [theme]
+  );
+
+  return (
+    <ThemeContext.Provider value={value}>{children}</ThemeContext.Provider>
+  );
+}
+
+export function useTheme() {
+  const context = useContext(ThemeContext);
+  if (context === undefined) {
+    throw new Error("useTheme은 ThemeProvider 내부에서 사용해야 합니다.");
+  }
+  return context;
+}
